Move shifter_file view tests to their own files
